fix: apply code-review findings to the generic-runtime refactor
Correctness fixes from the xhigh /code-review pass (12 of 15 findings):

- run: wrap every worker thread in _guarded_worker — an uncaught worker
  exception now logs, sets stop_event, and re-raises from run(), instead
  of hanging the pipeline or exiting 0 with nothing written
- run: _validate_inversion fails fast on invert=True with a
  non-invertible preprocess and on unrecognized inverse_stage strings;
  writer defaults a missing inverse_stage to after_finalize
- run/workers: gate deepcopy/.to()/.eval() on isinstance(nn.Module) so
  plain-callable BlockProcessors work as the protocol promises
- workflow: execution is now Optional (None -> synthesized from cfg,
  same rule as preprocess), so CLI --compile/--no-amp apply under
  --workflow unless the recipe pins its own policy
- execution/workers: cuda_safe_compile_mode mirrors the config-level
  CUDA-graph downgrade for directly-injected ExecutionPolicies
- run: reject output_spec_factory workflows without stores and
  OutputSpec(store=None) up front, not in a writer thread
- cli: build the workflow before opening --out-spec stores; reject
  --out-spec when the workflow has fixed outputs, require it otherwise;
  reject --weights with --workflow instead of silently ignoring it
- accumulators: MajorityVote skips NaN labels (no per-add vote-plane
  leak), canonicalizes float keys, and finalizes zero-vote voxels to
  background 0; document float-precision limits for instance IDs
- workers/accumulators: drop count/total from the BlockAccumulator
  protocol — the writer tracks block completion itself (_BlockState)
- workers/context: BlockContext is built once in prep and carried on
  Batch/Preds; delete from_preds and the full_shape threading
- outputs: Threshold uses float32 scalars (avoids a float64 temporary)
- transforms: document the before_accumulate block-level-ctx limitation

Skipped (need design work): integer-ID dtype path through the float
pipeline, per-patch ctx for before_accumulate inverses, and decoupling
config trim/blend validation from injected accumulators.

+++ src/aind_torch_utils/outputs.py
@@ -65,9 +65,11 @@ class Threshold:
         self.below = float(below)
 
     def __call__(self, block: np.ndarray, ctx: "BlockContext") -> np.ndarray:
-        return np.where(block > self.thresh, self.above, self.below).astype(
-            np.float32, copy=False
-        )
+        # float32 branch scalars keep np.where from promoting the whole block
+        # to float64 (Python-float branches would double the allocation).
+        return np.where(
+            block > self.thresh, np.float32(self.above), np.float32(self.below)
+        ).astype(np.float32, copy=False)
 
 
 class ThresholdThenOpen:
